Now with moveable camera and delete support
The camera system still has to be seriously fixed.
I'm just throwing a quick hack with sliders for the
first release.

There's a still an annoying but when creating a new
emotion and editing the previous one. When doing this,
the previous emotion is like overwritten by the new
one when editing it...
I still have to idea where this problem comes from.

Anyway, one bug to resolve before releasing the first
alpha.

Signed-off-by: VR Voyage <avatarsetup_voyage@miouyouyou.fr>