0.0.1rc1 : First release !
Yay ! The thing looks like a weird student card creator, is
extremely clunky but it's my clunky little tool, and I like
it, because it works for what I designed it.

So, yeah, unless a major bug pops up, this will be the first
release of the tool.
More to come soon.

Signed-off-by: VR Voyage <avatarsetup_voyage@miouyouyou.fr>