GLTF support: Not broken edition
Oh boy... PackedSceneGLTF is just BROKEN, even on recent versions
of Godot.

For some reasons, some geniuses in the Godot team thought it would be
INTELLIGENT TO REMOVE THE FEATURE AFTER EXPORTING THE EXE.

Meaning that PackedSceneGLTF is present and works very well when
in play mode, but suddenly BREAKS after exporting the project.

The VRM team thought about that and actually provides a DLL providing
the same class (because, the exporter can't do that, for SOME REASONS)
with different functions signatures.
Yes, different functions signatures, meaning that you have to add some
code that goes like "Is it the real one, or the DLL kludge ? If that's
the kludge, I need to call the function like this, if that's the one
included, I need to call the function like that...".

WHY Godot, WHY THE FUCK is this feature still not implemented correctly
in 2022 !? All the code is ALREADY THERE, and people are able to
generate DLL PROVIDING the feature on export.

Ugh...

Signed-off-by: Voyage <voyage@miouyouyou.fr>